Add `as` option for `this.mount` from RouterDSL.
The value of `options.as` will be used primarily, and will fallback to
the name of the engine itself.

Also ensures that the engine name is reset when using `as`. When using
`as` inside a nested route, linking into the engine requires only the
`as` value.  Without this change, the link added here would be
`{{#link-to 'routeable-engine-demo.blog.new'` instead of `{{#link-to 'blog.new'`.