Ensure engine rendered views/components have lifecycle hooks called.
If we do not share the parent `env`'s `lifecycleHooks` array, then we
will never have `didInsertElement`, `willDestroy`, etc hooks called.
If `didInsertElement` is not called, the view is not registered with the
`_viewRegistry` which means that no actions can be dispatched (since the
app/global EventDispatcher will not be aware of the engines views).