Install test addons as devDependencies.
The alternative, to specify paths in `ember-addon`, results in those
paths being loaded by consumers of this addon, which we don’t want.
Ultimately, it would be best to have a way to specify `devPaths` which
are loaded only for development of this addon, but not by its
consumers.

Note that symlinks are configured in the `prepublish` task so that
changes to the test addons will be picked up during dev.

[Fixes #43]