Remove usage of EmptyObject.
`EmptyObject` in Ember is an attempt to make creating
`Object.create(null)` objects faster by avoiding the creation of an
object with a null prototype.

In our case, these objects are not being created in a hot path
repeatedly so the performance concern is not valid. Using
`Object.create(null)` directly reduces our coupling to changes in Ember
and embraces "the platform" (:troll:).