-Update so that addons can accept asset POJOs containing environment-specific properties, instead of just string paths.
Context:
In our addon, we are importing a vendor file and specifying both a development and production (minified) version of the file. I received an error when trying to use the environment object version of the asset argument, which I traced down to the `asset.replace` (which assumes a string).