Prevent hoisting of all `app.import`'s.
Before this change, every addon that did an `app.import` on the
argument passed in to the `included` hook would end up hoisting
the thing imported all the way to the top.

This change passes in the engine itself as the `app` so that it is
properly handled (addons downstream from a lazy engine end up in the
lazy engines `engine-vendor.{js,css}` and addons downstream from
eager engines end up in the top level `vendor.{js,css}`).