Only include app/styles from dependency addons in lazy case
- Currently, engines, which are acting as addons (not lazy engines),
will build their own styles (everything in their addon/styles +
dependency addon/styles + app/styles). When acting as lazy engines,
however, it is presumed that the engine needs to build the
app-namespaced styles as well (similar to what the application will do,
I think). However, they are already built as part of the addon styles.
- Changing the code to only include the app/styles from the dependencies
in case of a lazy engine, ensures they are not duplicated in both
engine.css and engine-vendor.css.