add public API for reseting the `memoize` cache
I encountered the need for this change when I tried writing a test for
an addon that interacts with ember-engines. The cache holds on to
broccoli trees, including their location on disk. I was using
[broccoli-test-helper][bth] which creates temp directories for each
test. Ember-engines would continue to refer to the previous test's
temp directories and fail in confusing ways. This new API allows me
to reset the caches between tests.

[bth]:https://github.com/broccolijs/broccoli-test-helper