change memoize to allow separate caches per ember-cli project
I encountered the need for this change when I tried writing a test for
an addon that interacts with ember-engines. The cache holds on to
broccoli trees, including their location on disk. I was using
broccoli-test-helper which creates temp directories for each
test. Ember-engines would continue to refer to the previous test's
temp directories and fail in confusing ways. After this change, the cache
is stored on the addon project so the function calls occur once
per build.