fix: findHost and findHostHost should allow undefined lazyLoading config
Normally engine's parent is the project and the findHost loop stops at
the first level dependency engine. If an engine's parent is another
engine it's also fine. However if the parent is an addon that did not
specify lazyLoading the build breaks saying cannot read enabled of
undefined.  This change allows addon uses engines for its dummy app.