Ensure _externalRoutes is defined on EngineInstance instances.
Prior to Ember 3.24 within a `setupRenderingTest` a link-to-external we
short curcuit well before attempting to invoke `_getExternalRoute`
(because no router was present, so it did nothing). As of Ember 3.24+
the router is setup and available on demand (including in these types of
rendering tests).

This moves the initial definition of `engineOwner._externalRoutes` to
`init` of the engine instance instance (so that we know it is always
initialized) and updates a couple of tests to ensure that a proper
external route mapping is setup.