Revert usage of `@embroider/macros`
At the moment, using `@embroider/macros` prevents applications from
leveraging parallelized builds. Since we were only using
`dependencySatisfies` we can replace that with `gte` from
`ember-compatibility-helpers` which _is_ parallelizable to avoid the
issue.

Note: this is not a great long term solution (we really should be
converging on usage of `@embroider/macros`), but unblocks things for
now.