[BUGFIX] Fallback to LinkTo behavior when outside an engine
Prior to Ember 3.24, LinkToExternal only modifier the `routeKey` when inside an engine; that is, when `owner.mountPoint` exists. With the new class that exists for 3.24+, this behavior regressed. Now, `_namespaceRoute` is overridden in all cases, which means that if `LinkToExternal` ever gets rendered outside of an engine, it will automatically fail since `owner._getExternalRoute` will not find the specified route.

Why this matters: Routing code that may be invoked by an engine does not necessarily live in the same addon as the engine itself. If we extract code that is intended to link to external routes, we still need to be able to test it in those addons, which don't have engines of their own. Prior to 3.24, this worked fine thanks to the `owner.mountPoint` check. By adding a similar check here, we ensure that `LinkToExternal` will make a best effort to find the correct route even if it's rendered outside an engine, such as in a testing suite.